chore(tests): move tests into source, but ignore them for the build step
