+++ .babelrc
@@ -12,5 +12,9 @@
       "target": "react-transform-catch-errors",
       "imports": ["react", "redbox-react"]
     }]
-  }
+  },
+  "ignore": [
+    "__tests__/*"
+  ]
+
 }
