+++ package.json
@@ -28,7 +28,7 @@
     "build": "babel --out-dir='lib' src",
     "prepublish": "npm run build",
     "postpublish": "./scripts/publish_gh_pages.sh",
-    "test": "babel --out-dir='lib' src && cd tests && npm install && ln -snf ../lib lib && cd .. && jest tests/"
+    "test": "jest src/"
   },
   "devDependencies": {
     "babel": "^5.8.23",
@@ -40,10 +40,12 @@
     "eslint": "^1.3.1",
     "eslint-plugin-react": "^3.3.1",
     "jest-cli": "^0.5.1",
+    "react": "^0.13.3",
     "react-hot-loader": "^1.3.0",
     "react-transform-catch-errors": "^0.1.1",
     "react-transform-webpack-hmr": "^0.1.4",
     "redbox-react": "^1.0.2",
+    "underscore": "^1.8.3",
     "webpack": "^1.12.1",
     "webpack-dev-middleware": "^1.2.0",
     "webpack-dev-server": "^1.10.1",
@@ -51,19 +53,13 @@
   },
   "jest": {
     "unmockedModulePathPatterns": [
-      "react",
-      "underscore"
+      "<rootDir>/node_modules/react",
+      "underscore",
+      "../utils/helpers",
+      "../utils/union-class-names",
+      "../utils/is-component-of-type"
     ],
-    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-    "testFileExtensions": [
-      "jsx",
-      "js"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "jsx"
-    ]
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest"
   },
   "license": "MIT",
   "dependencies": {
