--- tests/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "belle",
-  "version": "1.2.2",
-  "description": "Tests for Belle",
-  "author": {
-    "name": "Nik Graf",
-    "email": "nik@nikgraf.com",
-    "url": "http://www.nikgraf.com"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "react": ">=0.13.0",
-    "underscore": ">=1.8.2"
-  },
-  "license": "MIT"
-}
